% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{current_value}
\alias{current_value}
\title{extract numeric value from ccurrent row (remove units)}
\usage{
current_value(df_col)
}
\arguments{
\item{df_col}{column (array) with data value and units}
}
\value{
extracted numeric value
}
\description{
extract numeric value from ccurrent row (remove units)
}
\examples{
\dontrun{current_value(dat$Sample)}
}
\author{
K. Juraic
}
