% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_fluke_csv}
\alias{read_fluke_csv}
\title{read Fluke multimeter CSV file}
\usage{
read_fluke_csv(file_name = file.choose())
}
\arguments{
\item{file_name}{name of CSV file with measured data}
}
\value{
data.frame() with read data
}
\description{
read Fluke multimeter CSV file
}
\examples{
\dontrun{read_fluke_csv("anodizacija_dat.csv")}
}
\author{
K. Juraic
}
